producer_consumer_app/pyproject.toml
[project]
name = "producer-consumer-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "kafka-python>=2.2.15",
    "prometheus-client>=0.22.1",
    "protobuf>=6.31.1",
    "pyyaml>=6.0.2",
]
